#649 - Learning curve diagram
- Separate class for Learning Curve Panel

- Fix error when there are no evaluations
