+++ inception-recommendation/src/main/java/de/tudarmstadt/ukp/inception/recommendation/sidebar/RecommendationSidebar.html
@@ -59,7 +59,7 @@
             <h3 class="panel-title"><wicket:message key="learningCurve"/></h3>
           </div>
           <div class="panel-body">
-            <div wicket:id="chart-container" ></div>
+            <div wicket:id="learningCurve" ></div>
           </div>
         </div>
       </div>
